requirements.txt
opencv-python
dill
numpy
matplotlib
pandas
scipy
pyyaml
h5py
scikit-image
numba
setuptools
scikit-learn
tqdm
torch
cudatoolkit==10.2